Removed the old InterfaceVariables
Removed the old interface variables file and renamed the new one to the appropriate name.
